cache.server.derivative.ttl_seconds is read as a long rather than an int

+++ src/main/java/edu/illinois/library/cantaloupe/cache/AzureStorageCache.java
@@ -136,7 +136,7 @@ class AzureStorageCache implements DerivativeCache {
 
     private static Instant getEarliestValidInstant() {
         final Configuration config = Configuration.getInstance();
-        final long ttl = config.getInt(Key.DERIVATIVE_CACHE_TTL);
+        final long ttl = config.getLong(Key.DERIVATIVE_CACHE_TTL);
         return (ttl > 0) ?
                 Instant.now().truncatedTo(ChronoUnit.SECONDS).minusSeconds(ttl) :
                 Instant.MIN;
