+++ src/main/java/edu/illinois/library/cantaloupe/cache/S3Cache.java
@@ -240,7 +240,7 @@ class S3Cache implements DerivativeCache {
      */
     private static Instant getEarliestValidInstant() {
         final Configuration config = Configuration.getInstance();
-        final long ttl = config.getInt(Key.DERIVATIVE_CACHE_TTL);
+        final long ttl = config.getLong(Key.DERIVATIVE_CACHE_TTL);
         return (ttl > 0) ?
                 Instant.now().truncatedTo(ChronoUnit.SECONDS).minusSeconds(ttl) :
                 Instant.EPOCH;
